feat(auth): add mock login and signup pages
